+++ app/src/App.vue
@@ -1,30 +1,14 @@
 <script setup>
-import { inject, ref } from 'vue'
-import { useRouter } from 'vue-router';
-
-import SvgIcon from '@jamescoyle/vue-icon';
+import { ref } from 'vue'
+import Header from './components/Header.vue';
 import SideBar from './components/SideBar.vue';
-import SocialStack from './components/SocialStack.vue';
-import { getUserInfo, getIconPath } from './utils';
+import { getUserInfo } from './utils';
 
-const resume = inject('resume');
-const isHome = ref(true);
-const sideBarWidth = ref("0px");
 const userLogin = ref(null);
-const router = useRouter();
+const sideBarWidth = ref("0px");
 
 function showSidebar() { sideBarWidth.value = "300px"};
 function collapseSidebar() { sideBarWidth.value = "0px"};
-
-router.beforeEach((to) => {
-    if (to.name == 'Home'){
-        isHome.value = true;
-    }
-    else{
-        isHome.value = false;
-        collapseSidebar();
-    }
-});
 
 getUserInfo().then(response => {
     userLogin.value = response
@@ -34,35 +18,9 @@ getUserInfo().then(response => {
 </script>
 
 <template>
-    <header>
-        <div class="links">
-            <RouterLink @click="showSidebar()" v-if="isHome" to=""><svg-icon type="mdi" :path="getIconPath('menu')" /></RouterLink>
-        </div>
-        <SocialStack :links="resume.links" />
-        <div class="userLogin">
-            <RouterLink v-if="!isHome" to="/"><svg-icon type="mdi" :path="getIconPath('home')" /></RouterLink>
-            <RouterLink v-else-if="isHome && !userLogin" to="/login"><svg-icon type="mdi" :path="getIconPath('loginvariant')" /></RouterLink>
-            <a v-else href="/.auth/logout"><svg-icon  type="mdi" :path="getIconPath('logoutvariant')" /></a>
-        </div>
-    </header>
-    <SideBar @collapse-sidebar="collapseSidebar()" :width="sideBarWidth" />
+    <Header @collapse-sidebar="collapseSidebar()" @show-sidebar="showSidebar()"></Header>
+    <SideBar @collapse-sidebar="collapseSidebar()" :width="sideBarWidth"></SideBar>
     <main>
         <router-view></router-view>
     </main>
 </template>
-
-<style scoped>
-.bottom-pad {
-    padding-bottom: 4.5rem;
-}
-
-.links {
-    display: flex;
-    justify-content: left;
-}
-
-.userLogin {
-    display: flex;
-    justify-content: right;
-}
-</style>
