+++ app/src/assets/main.scss
@@ -37,24 +37,15 @@ main {
 }
 
 header {
-    display: grid;
+    display: flex;
+    justify-content: space-between;
     top: 0;
     position: sticky;
     width: 100%;
-    grid-template-columns: 20% 60% 20%;
     padding: 1rem 2rem;
     box-shadow: var(--box-shadow-1);
     background-color: var(--vt-c-black);
     z-index: 1;
-}
-
-footer {
-    width: 100%;
-    position: fixed;
-    bottom: 0rem;
-    padding: 1rem 2rem;
-    box-shadow: var(--box-shadow-footer);
-    background-color: var(--vt-c-black);
 }
 
 button {
@@ -166,7 +157,7 @@ b {
     padding: 0rem 0.25rem;
     justify-content: space-between;
     align-items: center;
-    background-color: hsla(0, 0, 0, 0.25);
+    background-color: hsla(0, 0%, 0%, 0.25);
     border-radius: 50px;
 
     .c-info-content {
@@ -183,6 +174,20 @@ b {
 
     p {
         font-size: 15px;
+    }
+}
+
+.NotFound, .Unauthorized {
+    display: flex;
+    text-align: center;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 80vh;
+  
+    button {
+      margin: 0.75rem;
+      width: 8rem;
     }
 }
