+++ app/src/main.js
@@ -8,6 +8,7 @@ import App from './App.vue'
 import AdminPage from './views/AdminPage.vue'
 import HomePage from './views/HomePage.vue'
 import NotFoundPage from './views/NotFoundPage.vue'
+import Unauthorized from './views/Unauthorized.vue'
 import LoginPage from './views/LoginPage.vue'
 
 const app = createApp(App);
@@ -18,6 +19,7 @@ const router = createRouter({
         { path: "/", name: "Home", component: HomePage},
         { path: "/admin", name: "Admin", component: AdminPage},
         { path: "/login", name: "Login", component: LoginPage},
+        { path: "/unauthorized", name: "Unauthorized", component: Unauthorized},
         { path: "/:notFound", name: "NotFound", component: NotFoundPage},
     ]
 });
