+++ app/src/utils.js
@@ -1,5 +1,6 @@
 import MD5 from "crypto-js/md5";
 import { mdiBookEducation, mdiHome, mdiLoginVariant, mdiLogoutVariant, mdiMenu, mdiKey, mdiHeart, mdiHeartOutline, mdiBriefcase, mdiAccount, mdiEmail, mdiFilePdfBox, mdiLinkedin, mdiGithub, mdiMicrosoft, mdiApplication, mdiDatabase, mdiDotNet, mdiGit, mdiIntegratedCircuitChip, mdiLanguageC, mdiLanguageCpp, mdiLanguageCsharp, mdiLanguageCss3, mdiLanguageHtml5, mdiLanguageJava, mdiLanguageJavascript, mdiLanguagePython, mdiMathIntegralBox, mdiMicrosoftAzure, mdiMicrosoftAzureDevops, mdiMicrosoftVisualStudio, mdiMicrosoftVisualStudioCode, mdiSalesforce, mdiSass, mdiVuejs, mdiTools, mdiFileCertificate, mdiChevronRight, mdiChevronLeft } from '@mdi/js'
+
 const monthNames = ["JAN", "FEB", "MAR", "APR", "MAY", "JUNE", "JULY", "AUG", "SEPT", "OCT", "NOV", "DEC"];
 
 export function displayAddress(obj) {
