+++ app/src/views/UnAuthorized.vue
@@ -4,18 +4,16 @@ import { mdiHome } from '@mdi/js';
 import { useRouter } from 'vue-router';
 
 const router = useRouter();
-
 function goToHomePage(){
     router.push({ name: "Home" });
 };
 </script>
-
 <template>
-    <div class="NotFound">
-        <h1 class="h1-custom">404: Not Found</h1>
+    <div class="Unauthorized">
+        <h1 class="h1-custom">401: Unauthorized</h1>
         <div>
-            <h3>The requested route was not found.</h3>
-            <h3>Go home you are drunk.</h3>
+            <h3>You don't have permissions for this page. Please check your login info and permissions and try again.</h3>
+            <h3>Go home.</h3>
         </div>
         <button @click="goToHomePage()">
             <svg-icon type="mdi" :path="mdiHome" :size="25"></svg-icon>
@@ -23,18 +21,3 @@ function goToHomePage(){
         </button>
     </div>
 </template>
-
-<style lang="scss" scoped>
-.NotFound {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 80vh;
-
-  button {
-    margin: 1rem;
-    width: 8rem;
-  }
-}
-</style>
