+++ app/staticwebapp.config.json
@@ -9,6 +9,12 @@
       "redirect": "/"
     }
   ],
+  "responseOverrides": {
+    "401": {
+      "statusCode": 401,
+      "redirect": "/unauthorized"
+    }
+  },
   "navigationFallback": {
     "rewrite": "/index.html",
     "exclude": ["*.{css, svg, js, ts, png, jpeg, jpg, ico}"]
